requirements.txt
ultralytics
pandas
pyyaml